feat: add game logic
When the user click the "play" button, a random list of words is generated. The first word is then played and the application waits for the correct card to be pressed. If true, then the next word is played, if an error, then the application will wait for the correct answer. When you click on the "repeat" button, the last word will be played.